aries/amla/python-app2/requirements.txt
mysql-connector-python
flask
requests
